% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R, R/mcmc.R
\name{match2sided}
\alias{match2sided}
\alias{joint_lpdf}
\alias{match2sided}
\title{Two sided matching model}
\usage{
joint_lpdf(opp, choice, alpha, beta, mu_beta, Tau_beta, prior, ww, xx)

match2sided(iter, eps_alpha, eps_beta, frac_beta, frac_opp, ww, xx, choice, opp)
}
\arguments{
\item{opp}{A matrix of opportunity}

\item{choice}{A vector of observed choice}

\item{alpha}{Multinomial preference}

\item{beta}{Binomial preference}

\item{mu_beta}{Hyperparameter of beta}

\item{prior}{A list of prior parameters}

\item{sigma_beta}{Hyperparameter of beta}
}
\description{
Two sided matching model

Calculate the log joint pdf, useful for testing

Run the MCMC estimation
}
